final/ml-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
joblib==1.4.2
scikit-learn==1.5.2
starlette-exporter==0.23.0
psycopg2-binary==2.9.9
